0.0.4
Time: 2019/12/4 23:55
Author: Gong Haodong & Yeti
Description: Add Activity and Fragment for getting DOM from given URL, but have no proper way to use them now.  The purpose for this uploading is to show the current MVP guideline followed.(based on google official example TODO-app),not all file are important now, only attention to the /chooseWatchingTarget files and the src/Base*.java files is needed.